+++ src/graphql/network.rs
@@ -13,7 +13,7 @@ use async_graphql::{
 };
 use chrono::{DateTime, Utc};
 use giganto_client::ingest::network::{
-    Conn, DceRpc, Dns, Ftp, Http, Kerberos, Ldap, Mqtt, Ntlm, Rdp, Smtp, Ssh,
+    Conn, DceRpc, Dns, Ftp, Http, Kerberos, Ldap, Mqtt, Ntlm, Rdp, Smtp, Ssh, Tls,
 };
 use serde::Serialize;
 use std::{collections::BTreeSet, fmt::Debug, iter::Peekable, net::IpAddr};
@@ -364,6 +364,35 @@ struct LdapRawEvent {
     argument: Vec<String>,
 }
 
+#[derive(SimpleObject, Debug)]
+struct TlsRawEvent {
+    timestamp: DateTime<Utc>,
+    orig_addr: String,
+    orig_port: u16,
+    resp_addr: String,
+    resp_port: u16,
+    proto: u8,
+    last_time: i64,
+    server_name: String,
+    alpn_protocol: String,
+    ja3: String,
+    version: String,
+    cipher: u16,
+    ja3s: String,
+    serial: String,
+    subject_country: String,
+    subject_org_name: String,
+    subject_common_name: String,
+    validity_not_before: i64,
+    validity_not_after: i64,
+    subject_alt_name: String,
+    issuer_country: String,
+    issuer_org_name: String,
+    issuer_org_unit_name: String,
+    issuer_common_name: String,
+    last_alert: u8,
+}
+
 #[allow(clippy::enum_variant_names)]
 #[derive(Union)]
 enum NetworkRawEvents {
@@ -378,6 +407,7 @@ enum NetworkRawEvents {
     FtpRawEvent(FtpRawEvent),
     MqttRawEvent(MqttRawEvent),
     LdapRawEvent(LdapRawEvent),
+    TlsRawEvent(TlsRawEvent),
 }
 
 macro_rules! from_key_value {
@@ -557,6 +587,29 @@ from_key_value!(
     diagnostic_message,
     object,
     argument
+);
+
+from_key_value!(
+    TlsRawEvent,
+    Tls,
+    server_name,
+    alpn_protocol,
+    ja3,
+    version,
+    cipher,
+    ja3s,
+    serial,
+    subject_country,
+    subject_org_name,
+    subject_common_name,
+    validity_not_before,
+    validity_not_after,
+    subject_alt_name,
+    issuer_country,
+    issuer_org_name,
+    issuer_org_unit_name,
+    issuer_common_name,
+    last_alert
 );
 
 #[Object]
@@ -861,6 +914,31 @@ impl NetworkQuery {
         .await
     }
 
+    async fn tls_raw_events<'ctx>(
+        &self,
+        ctx: &Context<'ctx>,
+        filter: NetworkFilter,
+        after: Option<String>,
+        before: Option<String>,
+        first: Option<i32>,
+        last: Option<i32>,
+    ) -> Result<Connection<String, TlsRawEvent>> {
+        let db = ctx.data::<Database>()?;
+        let store = db.tls_store()?;
+        let key_prefix = key_prefix(&filter.source);
+
+        query(
+            after,
+            before,
+            first,
+            last,
+            |after, before, first, last| async move {
+                load_connection(&store, &key_prefix, &filter, after, before, first, last)
+            },
+        )
+        .await
+    }
+
     #[allow(clippy::too_many_lines)]
     async fn network_raw_events<'ctx>(
         &self,
@@ -1077,6 +1155,24 @@ impl NetworkQuery {
                     }
                 }
 
+                let (tls_iter, cursor, _) = get_filtered_iter(
+                    &db.tls_store()?,
+                    &key_prefix,
+                    &filter,
+                    &after,
+                    &before,
+                    first,
+                    last,
+                )?;
+                let mut tls_iter = tls_iter.peekable();
+                if let Some(cursor) = cursor {
+                    if let Some((key, _)) = tls_iter.peek() {
+                        if key.as_ref() == cursor {
+                            tls_iter.next();
+                        }
+                    }
+                }
+
                 let mut is_forward: bool = true;
                 if before.is_some() || last.is_some() {
                     is_forward = false;
@@ -1094,6 +1190,7 @@ impl NetworkQuery {
                     ftp_iter,
                     mqtt_iter,
                     ldap_iter,
+                    tls_iter,
                     size,
                     is_forward,
                 )
@@ -1269,6 +1366,20 @@ impl NetworkQuery {
             .collect::<BTreeSet<(DateTime<Utc>, Vec<u8>)>>();
         Ok(collect_exist_timestamp::<Ldap>(&exist_data, &filter))
     }
+
+    async fn search_tls_raw_events<'ctx>(
+        &self,
+        ctx: &Context<'ctx>,
+        filter: SearchFilter,
+    ) -> Result<Vec<DateTime<Utc>>> {
+        let db = ctx.data::<Database>()?;
+        let store = db.tls_store()?;
+        let exist_data = store
+            .multi_get_from_ts(&filter.source, &filter.timestamps)
+            .into_iter()
+            .collect::<BTreeSet<(DateTime<Utc>, Vec<u8>)>>();
+        Ok(collect_exist_timestamp::<Tls>(&exist_data, &filter))
+    }
 }
 
 #[allow(clippy::too_many_arguments, clippy::too_many_lines)]
@@ -1284,6 +1395,7 @@ fn network_connection(
     mut ftp_iter: Peekable<FilteredIter<Ftp>>,
     mut mqtt_iter: Peekable<FilteredIter<Mqtt>>,
     mut ldap_iter: Peekable<FilteredIter<Ldap>>,
+    mut tls_iter: Peekable<FilteredIter<Tls>>,
     size: usize,
     is_forward: bool,
 ) -> Result<Connection<String, NetworkRawEvents>> {
@@ -1304,6 +1416,7 @@ fn network_connection(
     let mut ftp_data = ftp_iter.next();
     let mut mqtt_data = mqtt_iter.next();
     let mut ldap_data = ldap_iter.next();
+    let mut tls_data = tls_iter.next();
 
     loop {
         let conn_ts = if let Some((ref key, _)) = conn_data {
@@ -1372,15 +1485,26 @@ fn network_connection(
             min_max_time(is_forward)
         };
 
-        let selected = if is_forward {
-            timestamp.min(dns_ts.min(conn_ts.min(http_ts.min(rdp_ts.min(ntlm_ts.min(
-                kerberos_ts.min(ssh_ts.min(dce_rpc_ts.min(ftp_ts.min(mqtt_ts.min(ldap_ts))))),
-            ))))))
+        let tls_ts = if let Some((ref key, _)) = tls_data {
+            get_timestamp(key)?
         } else {
-            timestamp.max(dns_ts.max(conn_ts.max(http_ts.max(rdp_ts.max(ntlm_ts.max(
-                kerberos_ts.max(ssh_ts.max(dce_rpc_ts.max(ftp_ts.max(mqtt_ts.max(ldap_ts))))),
-            ))))))
+            min_max_time(is_forward)
         };
+
+        let selected =
+            if is_forward {
+                timestamp.min(dns_ts.min(conn_ts.min(http_ts.min(rdp_ts.min(ntlm_ts.min(
+                    kerberos_ts.min(
+                        ssh_ts.min(dce_rpc_ts.min(ftp_ts.min(mqtt_ts.min(ldap_ts.min(tls_ts))))),
+                    ),
+                ))))))
+            } else {
+                timestamp.max(dns_ts.max(conn_ts.max(http_ts.max(rdp_ts.max(ntlm_ts.max(
+                    kerberos_ts.max(
+                        ssh_ts.max(dce_rpc_ts.max(ftp_ts.max(mqtt_ts.max(ldap_ts.max(tls_ts))))),
+                    ),
+                ))))))
+            };
 
         match selected {
             _ if selected == conn_ts => {
@@ -1497,6 +1621,16 @@ fn network_connection(
                 } else {
                 };
             }
+            _ if selected == tls_ts => {
+                if let Some((key, value)) = tls_data {
+                    result_vec.push(Edge::new(
+                        base64_engine.encode(&key),
+                        NetworkRawEvents::TlsRawEvent(TlsRawEvent::from_key_value(&key, value)?),
+                    ));
+                    tls_data = tls_iter.next();
+                } else {
+                };
+            }
             _ => {}
         }
         if (result_vec.len() >= size)
@@ -1510,7 +1644,8 @@ fn network_connection(
                 && dce_rpc_data.is_none()
                 && ftp_data.is_none()
                 && mqtt_data.is_none()
-                && ldap_data.is_none())
+                && ldap_data.is_none()
+                && tls_data.is_none())
         {
             if conn_data.is_some()
                 || dns_data.is_some()
@@ -1523,6 +1658,7 @@ fn network_connection(
                 || ftp_data.is_some()
                 || mqtt_data.is_some()
                 || ldap_data.is_some()
+                || tls_data.is_some()
             {
                 has_next_value = true;
             }
@@ -1563,7 +1699,7 @@ mod tests {
     use crate::storage::RawEventStore;
     use chrono::{Duration, TimeZone, Utc};
     use giganto_client::ingest::network::{
-        Conn, DceRpc, Dns, Ftp, Http, Kerberos, Ldap, Mqtt, Ntlm, Rdp, Smtp, Ssh,
+        Conn, DceRpc, Dns, Ftp, Http, Kerberos, Ldap, Mqtt, Ntlm, Rdp, Smtp, Ssh, Tls,
     };
     use std::mem;
     use std::net::IpAddr;
@@ -2407,6 +2543,73 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn tls_with_data() {
+        let schema = TestSchema::new();
+        let store = schema.db.tls_store().unwrap();
+
+        insert_tls_raw_event(&store, "src 1", Utc::now().timestamp_nanos());
+        insert_tls_raw_event(&store, "src 1", Utc::now().timestamp_nanos());
+
+        let query = r#"
+        {
+            tlsRawEvents(
+                filter: {
+                    source: "src 1"
+                }
+                first: 1
+            ) {
+                edges {
+                    node {
+                        origAddr,
+                    }
+                }
+            }
+        }"#;
+        let res = schema.execute(query).await;
+        assert_eq!(
+            res.data.to_string(),
+            "{tlsRawEvents: {edges: [{node: {origAddr: \"192.168.4.76\"}}]}}"
+        );
+    }
+
+    fn insert_tls_raw_event(store: &RawEventStore<Tls>, source: &str, timestamp: i64) {
+        let mut key = Vec::with_capacity(source.len() + 1 + mem::size_of::<i64>());
+        key.extend_from_slice(source.as_bytes());
+        key.push(0);
+        key.extend(timestamp.to_be_bytes());
+
+        let tls_body = Tls {
+            orig_addr: "192.168.4.76".parse::<IpAddr>().unwrap(),
+            orig_port: 46378,
+            resp_addr: "31.3.245.133".parse::<IpAddr>().unwrap(),
+            resp_port: 80,
+            proto: 17,
+            last_time: 1,
+            server_name: "server_name".to_string(),
+            alpn_protocol: "alpn_protocol".to_string(),
+            ja3: "ja3".to_string(),
+            version: "version".to_string(),
+            cipher: 10,
+            ja3s: "ja3s".to_string(),
+            serial: "serial".to_string(),
+            subject_country: "sub_contry".to_string(),
+            subject_org_name: "sub_org".to_string(),
+            subject_common_name: "sub_comm".to_string(),
+            validity_not_before: 11,
+            validity_not_after: 12,
+            subject_alt_name: "sub_alt".to_string(),
+            issuer_country: "issuer_contry".to_string(),
+            issuer_org_name: "issuer_org".to_string(),
+            issuer_org_unit_name: "issuer_org_unit".to_string(),
+            issuer_common_name: "issuer_comm".to_string(),
+            last_alert: 13,
+        };
+        let ser_tls_body = bincode::serialize(&tls_body).unwrap();
+
+        store.append(&key, &ser_tls_body).unwrap();
+    }
+
+    #[tokio::test]
     async fn conn_with_start_or_end() {
         let schema = TestSchema::new();
         let store = schema.db.conn_store().unwrap();
@@ -2456,6 +2659,7 @@ mod tests {
         let ftp_store = schema.db.ftp_store().unwrap();
         let mqtt_store = schema.db.mqtt_store().unwrap();
         let ldap_store = schema.db.ldap_store().unwrap();
+        let tls_store = schema.db.tls_store().unwrap();
 
         insert_conn_raw_event(
             &conn_store,
@@ -2534,8 +2738,15 @@ mod tests {
                 .unwrap()
                 .timestamp_nanos(),
         );
+        insert_tls_raw_event(
+            &tls_store,
+            "src 1",
+            Utc.with_ymd_and_hms(2023, 1, 6, 11, 11, 0)
+                .unwrap()
+                .timestamp_nanos(),
+        );
 
-        // order: ssh, conn, rdp, dce_rpc, http, dns, ntlm, kerberos, ftp, mqtt, ldap
+        // order: ssh, conn, rdp, dce_rpc, http, dns, ntlm, kerberos, ftp, mqtt,tls, ldap
         let query = r#"
         {
             networkRawEvents(
@@ -2580,13 +2791,16 @@ mod tests {
                         ... on LdapRawEvent {
                             timestamp
                         }
+                        ... on TlsRawEvent {
+                            timestamp
+                        }
                         __typename
                     }
                 }
             }
         }"#;
         let res = schema.execute(query).await;
-        assert_eq!(res.data.to_string(), "{networkRawEvents: {edges: [{node: {timestamp: \"2020-01-01T00:00:01+00:00\",__typename: \"SshRawEvent\"}},{node: {timestamp: \"2020-01-01T00:01:01+00:00\",__typename: \"ConnRawEvent\"}},{node: {timestamp: \"2020-01-05T00:01:01+00:00\",__typename: \"RdpRawEvent\"}},{node: {timestamp: \"2020-01-05T06:05:00+00:00\",__typename: \"DceRpcRawEvent\"}},{node: {timestamp: \"2020-06-01T00:01:01+00:00\",__typename: \"HttpRawEvent\"}},{node: {timestamp: \"2021-01-01T00:01:01+00:00\",__typename: \"DnsRawEvent\"}},{node: {timestamp: \"2022-01-05T00:01:01+00:00\",__typename: \"NtlmRawEvent\"}},{node: {timestamp: \"2023-01-05T00:01:01+00:00\",__typename: \"KerberosRawEvent\"}},{node: {timestamp: \"2023-01-05T12:12:00+00:00\",__typename: \"FtpRawEvent\"}},{node: {timestamp: \"2023-01-05T12:12:00+00:00\",__typename: \"MqttRawEvent\"}},{node: {timestamp: \"2023-01-06T12:12:00+00:00\",__typename: \"LdapRawEvent\"}}]}}");
+        assert_eq!(res.data.to_string(), "{networkRawEvents: {edges: [{node: {timestamp: \"2020-01-01T00:00:01+00:00\",__typename: \"SshRawEvent\"}},{node: {timestamp: \"2020-01-01T00:01:01+00:00\",__typename: \"ConnRawEvent\"}},{node: {timestamp: \"2020-01-05T00:01:01+00:00\",__typename: \"RdpRawEvent\"}},{node: {timestamp: \"2020-01-05T06:05:00+00:00\",__typename: \"DceRpcRawEvent\"}},{node: {timestamp: \"2020-06-01T00:01:01+00:00\",__typename: \"HttpRawEvent\"}},{node: {timestamp: \"2021-01-01T00:01:01+00:00\",__typename: \"DnsRawEvent\"}},{node: {timestamp: \"2022-01-05T00:01:01+00:00\",__typename: \"NtlmRawEvent\"}},{node: {timestamp: \"2023-01-05T00:01:01+00:00\",__typename: \"KerberosRawEvent\"}},{node: {timestamp: \"2023-01-05T12:12:00+00:00\",__typename: \"FtpRawEvent\"}},{node: {timestamp: \"2023-01-05T12:12:00+00:00\",__typename: \"MqttRawEvent\"}},{node: {timestamp: \"2023-01-06T11:11:00+00:00\",__typename: \"TlsRawEvent\"}},{node: {timestamp: \"2023-01-06T12:12:00+00:00\",__typename: \"LdapRawEvent\"}}]}}");
     }
 
     #[tokio::test]
@@ -3039,6 +3253,42 @@ mod tests {
         assert_eq!(
             res.data.to_string(),
             "{searchLdapRawEvents: [\"2020-01-01T00:01:01+00:00\",\"2020-01-01T01:01:01+00:00\"]}"
+        );
+    }
+
+    #[tokio::test]
+    async fn search_tls_with_data() {
+        let schema = TestSchema::new();
+        let store = schema.db.tls_store().unwrap();
+
+        let timestamp1 = Utc.with_ymd_and_hms(2020, 1, 1, 0, 0, 1).unwrap(); //2020-01-01T00:00:01Z
+        let timestamp2 = Utc.with_ymd_and_hms(2020, 1, 1, 0, 1, 1).unwrap(); //2020-01-01T00:01:01Z
+        let timestamp3 = Utc.with_ymd_and_hms(2020, 1, 1, 1, 1, 1).unwrap(); //2020-01-01T01:01:01Z
+        let timestamp4 = Utc.with_ymd_and_hms(2020, 1, 2, 0, 0, 1).unwrap(); //2020-01-02T00:00:01Z
+
+        insert_tls_raw_event(&store, "src 1", timestamp1.timestamp_nanos());
+        insert_tls_raw_event(&store, "src 1", timestamp2.timestamp_nanos());
+        insert_tls_raw_event(&store, "src 1", timestamp3.timestamp_nanos());
+        insert_tls_raw_event(&store, "src 1", timestamp4.timestamp_nanos());
+
+        let query = r#"
+        {
+            searchTlsRawEvents(
+                filter: {
+                    time: { start: "2020-01-01T00:01:01Z", end: "2020-01-01T01:01:02Z" }
+                    source: "src 1"
+                    origAddr: { start: "192.168.4.75", end: "192.168.4.79" }
+                    respAddr: { start: "31.3.245.130", end: "31.3.245.135" }
+                    origPort: { start: 46377, end: 46380 }
+                    respPort: { start: 75, end: 85 }
+                    timestamps:["2020-01-01T00:00:01Z","2020-01-01T00:01:01Z","2020-01-01T01:01:01Z","2020-01-02T00:00:01Z"]
+                }
+            )
+        }"#;
+        let res = schema.execute(query).await;
+        assert_eq!(
+            res.data.to_string(),
+            "{searchTlsRawEvents: [\"2020-01-01T00:01:01+00:00\",\"2020-01-01T01:01:01+00:00\"]}"
         );
     }
 }